#438 Revit_oM Tidied - enums split into separate files
