+++ Revit_oM/Elements/Viewport.cs
@@ -1,6 +1,6 @@
 /*
  * This file is part of the Buildings and Habitats object Model (BHoM)
- * Copyright (c) 2015 - 2018, the respective contributors. All rights reserved.
+ * Copyright (c) 2015 - 2019, the respective contributors. All rights reserved.
  *
  * Each contributor holds copyright over their respective contributions.
  * The project versioning (Git) records all such contribution source information.
@@ -34,7 +34,7 @@ namespace BH.oM.Adapters.Revit.Elements
 
         public InstanceProperties InstanceProperties { get; set; } = new InstanceProperties();
 
-        public Point Location { get; set; } = new Point() {X =0, Y = 0, Z = 0 };
+        public Point Location { get; set; } = new Point() { X =0, Y = 0, Z = 0 };
 
         /***************************************************/
     }
